fix(ibkr): move IbkrSyncService into com.picsou.service
IbkrSyncService called AccountService.toResponse / upsertSnapshot, which are
package-private. Placing the service in com.picsou.ibkr made those calls fail to
compile across packages. Move it to com.picsou.service alongside the other broker
sync services (Trade Republic, Bourso), which is where it belongs and where that
package-private access is legal. Update IbkrController + the test package, drop the
now same-package import in SchedulerService, and fix the doc path.
